requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
PyPDF2>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
google-generativeai>=0.3.0
psutil>=5.9.0
pytest>=7.4.0 